object name, comments: update modified
It looks like we were setting the object name and comments for the current
image directly, which resulted in us not doing some bookkeeping. In particular,
we weren't always setting `self._modified`, which we expect to be set when
saving to the database, so that was throwing an error. I suspect that that
error was being silently handled, so it would look like things went okay, but
it would result in us failing to save data and doing whatever would normally
come after that.

Additionally, when clearing the object name and comments, we weren't keeping
track of modified state, and we were setting the values on the `ImcoImage`
instance to `None` instead of the empty string. I think we should be keeping
track of modified state because this *is* a change to the coding for the image,
and I think we should be using the empty string because that's how we
initialize the object name and comments in `__init__`.