remove object name, comments: stop saving session
This doesn't seem right to me, since we don't unconditionally save the session
when adding other codes to an image.

Also, move the object name and comments clearing logic to methods rather than
lambdas. This follows the pattern of other app actions, and it keeps the
"remove" logic next to the "set" logic.