change bindings to avoid conflicts
This may not be necessary, but I noticed that there were circumstances under
which hitting Command+N would toggle the N code (natural object). There are
other fixes coming that _may_ address that, but it seems like a good idea to
play it safe and make sure that one key event can't trigger two bindings.

Also, change "Cmd" to "Command" for the macOS menu accelerator "meta" key
because that's what the docs speficy, and it seems to work to make the
appropriate symbol show up in the menu.