macOS: don't bind left/right/enter
It appears that, on macOS, the menu accelerators that don't include a meta key
are automatically bound to the menu event handler. This means that if we define
a separate key binding attached to the same event handler, we get two event
handlers for the same event, resulting in all kinds of craziness.

This behavior would appear to happen only sometimes because of another bug that
was causing us to lose focus, after which the menu-accelerator bindings continue
to work while the manual bindings do not, resulting in just one event handler
per key press.

We also remove the <Enter> binding here for fear that it would be triggered by,
e.g., hitting <Enter> to dismiss a dialog. Let's play it safe.